+++ src/v1/BasicTypes/Objects/getProperty.ts
@@ -55,6 +55,6 @@
 export function getProperty<T = unknown>(
      target: object,
      propName: string|number|symbol
-): T {
+): T|undefined {
     return (target as Record<typeof propName, any>)[propName as any] ?? undefined;
 }