+++ src/v1/BasicTypes/HashMaps/HashMap.spec.ts
@@ -616,4 +616,29 @@ describe("HashMap()", () => {
             expect(actualResult).eql(expectedResult);
         });
     });
+
+    describe(".size()", () => {
+        it ("returns the number of keys in the given HashMap", () => {
+            // ----------------------------------------------------------------
+            // setup your test
+
+            const unit: HashMap<string> = {
+                attr1: "this is attr1",
+                attr2: "this is attr2",
+                attr3: "this is attr3"
+            }
+
+            const expectedResult = 3;
+
+            // ----------------------------------------------------------------
+            // perform the change
+
+            const actualResult = HashMap.size(unit);
+
+            // ----------------------------------------------------------------
+            // test the results
+
+            expect(actualResult).eql(expectedResult);
+        });
+    });
 });