+++ src/v1/BasicTypes/HashMaps/HashMap.spec.ts
@@ -33,6 +33,7 @@
 //
 import { expect } from "chai";
 import { describe } from "mocha";
+import { resolveNumerical } from "../../OptionTypes";
 
 import { HashMap } from "./HashMap";
 
@@ -810,4 +811,107 @@ describe("HashMap()", () => {
             expect(actualResult).eql(expectedResult);
         });
     });
+
+    describe(".map()", () => {
+        it ("calls the given callbackfn exactly once for each property", () => {
+            // ----------------------------------------------------------------
+            // setup your test
+
+            const unit: HashMap<string> = {
+                attr1: "this is attr1",
+                attr2: "this is attr2",
+                attr3: "this is attr3"
+            }
+
+            const expectedResult = {
+                attr1: 1,
+                attr2: 1,
+                attr3: 1,
+            };
+
+            // ----------------------------------------------------------------
+            // perform the change
+
+            const actualResult: HashMap<number> = {
+                attr1: 0,
+                attr2: 0,
+                attr3: 0,
+            }
+            HashMap.map(unit, (value, key) => {
+                actualResult[key]++;
+            })
+
+            // ----------------------------------------------------------------
+            // test the results
+
+            expect(actualResult).eql(expectedResult);
+        });
+
+        it ("returns a new object", () => {
+            // ----------------------------------------------------------------
+            // setup your test
+
+            const unit: HashMap<string> = {
+                attr1: "this is attr1",
+                attr2: "this is attr2",
+                attr3: "this is attr3"
+            }
+
+            const expectedResult = {
+                attr1: "this is attr1",
+                attr2: "this is attr2",
+                attr3: "this is attr3",
+                attr4: "this is not unit",
+            };
+
+            // ----------------------------------------------------------------
+            // perform the change
+
+            // this will build a new object that is identical to `unit`
+            const actualResult = HashMap.map(unit, (value) => {
+                return value;
+            });
+
+            // now, we need to modify `actualResult` to prove that it
+            // is different to `unit`
+            actualResult.attr4 = "this is not unit";
+
+            // ----------------------------------------------------------------
+            // test the results
+
+            expect(actualResult).eql(expectedResult);
+            expect(actualResult).not.eql(unit);
+        });
+
+        it ("the callbackfn can change the type of each value", () => {
+            // ----------------------------------------------------------------
+            // setup your test
+
+            const unit: HashMap<string|boolean> = {
+                attr1: "100",
+                attr2: true,
+                attr3: false,
+            }
+
+            const expectedResult: HashMap<number> = {
+                attr1: 100,
+                attr2: 1,
+                attr3: 0,
+            };
+
+            // ----------------------------------------------------------------
+            // perform the change
+
+            // this will build a new object that is identical to `unit`
+            const actualResult = HashMap.map(unit, (value) => {
+                return resolveNumerical(value);
+            });
+
+            // ----------------------------------------------------------------
+            // test the results
+
+            expect(actualResult).eql(expectedResult);
+        });
+    });
+
 });