+++ CHANGELOG.md
@@ -21,6 +21,11 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+### Fixes
+
+* `getProperty()` type signature updated to show that it can return `undefined`
+  - unfortunately, the TS compiler wasn't always able to work this out by itself
+
 ## v0.3.6
 
 Released Sunday, 1st May 2022.