+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The following changes have been completed, and will be included in the next tagg
 
 * Added `HashMap.hasProperty()`
 * Added `HashMap.getProperty()`
+* Added `HashMap.size()`
 
 ### Fixes