+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The following changes have been completed, and will be included in the next tagg
 * Added `HashMap.has()`
 * Added `HashMap.get()`
 * Added `HashMap.size()`
+* Added `deleteProperty()`
 
 ### Fixes