+++ CHANGELOG.md
@@ -21,6 +21,10 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+## v0.3.7
+
+Released Monday, 2nd May 2022.
+
 ### New
 
 The focus of this release has been on making HashMap much more complete.